requirements.txt
abutils
biopython
celery
nwalign3
pymongo
pytest
scikit-bio
